session, server: do simple refactor
