+++ session/bootstrap.go
@@ -29,11 +29,9 @@ import (
 	"time"
 
 	"github.com/pingcap/errors"
-	"github.com/pingcap/failpoint"
 	"github.com/pingcap/tidb/bindinfo"
 	"github.com/pingcap/tidb/config"
 	"github.com/pingcap/tidb/ddl"
-	"github.com/pingcap/tidb/ddl/syncer"
 	"github.com/pingcap/tidb/domain"
 	"github.com/pingcap/tidb/domain/infosync"
 	"github.com/pingcap/tidb/expression"
@@ -1256,113 +1254,6 @@ func upgrade(s Session) {
 			zap.Int64("to", currentBootstrapVersion),
 			zap.Error(err))
 	}
-}
-
-// SyncUpgradeState syncs upgrade state to etcd.
-func SyncUpgradeState(s Session) error {
-	totalInterval := time.Duration(internalSQLTimeout) * time.Second
-	ctx, cancelFunc := context.WithTimeout(context.Background(), totalInterval)
-	defer cancelFunc()
-	dom := domain.GetDomain(s)
-	err := dom.DDL().StateSyncer().UpdateGlobalState(ctx, syncer.NewStateInfo(syncer.StateUpgrading))
-	if err != nil {
-		logutil.BgLogger().Error("update global state failed", zap.String("category", "upgrading"), zap.String("state", syncer.StateUpgrading), zap.Error(err))
-		return err
-	}
-
-	interval := 200 * time.Millisecond
-	retryTimes := int(totalInterval / interval)
-	for i := 0; i < retryTimes; i++ {
-		op, err := owner.GetOwnerOpValue(ctx, dom.EtcdClient(), ddl.DDLOwnerKey, "upgrade bootstrap")
-		if err == nil && op.String() == owner.OpGetUpgradingState.String() {
-			break
-		}
-		if i == retryTimes-1 {
-			logutil.BgLogger().Error("get owner op failed", zap.String("category", "upgrading"), zap.Stringer("state", op), zap.Error(err))
-			return err
-		}
-		if i%10 == 0 {
-			logutil.BgLogger().Warn("get owner op failed", zap.String("category", "upgrading"), zap.Stringer("state", op), zap.Error(err))
-		}
-		time.Sleep(interval)
-	}
-
-	logutil.BgLogger().Info("update global state to upgrading", zap.String("category", "upgrading"), zap.String("state", syncer.StateUpgrading))
-	return nil
-}
-
-// SyncNormalRunning syncs normal state to etcd.
-func SyncNormalRunning(s Session) error {
-	failpoint.Inject("mockResumeAllJobsFailed", func(val failpoint.Value) {
-		if val.(bool) {
-			dom := domain.GetDomain(s)
-			//nolint: errcheck
-			dom.DDL().StateSyncer().UpdateGlobalState(context.Background(), syncer.NewStateInfo(syncer.StateNormalRunning))
-			failpoint.Return(nil)
-		}
-	})
-
-	jobErrs, err := ddl.ResumeAllJobsBySystem(s)
-	if err != nil {
-		logutil.BgLogger().Warn("resume all paused jobs failed", zap.String("category", "upgrading"), zap.Error(err))
-	}
-	for _, e := range jobErrs {
-		logutil.BgLogger().Warn("resume the job failed ", zap.String("category", "upgrading"), zap.Error(e))
-	}
-
-	ctx, cancelFunc := context.WithTimeout(context.Background(), 3*time.Second)
-	defer cancelFunc()
-	dom := domain.GetDomain(s)
-	err = dom.DDL().StateSyncer().UpdateGlobalState(ctx, syncer.NewStateInfo(syncer.StateNormalRunning))
-	if err != nil {
-		logutil.BgLogger().Error("update global state to normal failed", zap.String("category", "upgrading"), zap.Error(err))
-		return err
-	}
-	logutil.BgLogger().Info("update global state to normal running finished", zap.String("category", "upgrading"))
-	return nil
-}
-
-// IsUpgradingClusterState checks whether the global state is upgrading.
-func IsUpgradingClusterState(s Session) (bool, error) {
-	dom := domain.GetDomain(s)
-	ctx, cancelFunc := context.WithTimeout(context.Background(), 3*time.Second)
-	defer cancelFunc()
-	stateInfo, err := dom.DDL().StateSyncer().GetGlobalState(ctx)
-	if err != nil {
-		return false, err
-	}
-
-	return stateInfo.State == syncer.StateUpgrading, nil
-}
-
-func checkOrSyncUpgrade(s Session, ver int64) {
-	if ver < SupportUpgradeHTTPOpVer {
-		terror.MustNil(SyncUpgradeState(s))
-		return
-	}
-
-	interval := 200 * time.Millisecond
-	retryTimes := int(time.Duration(internalSQLTimeout) * time.Second / interval)
-	for i := 0; i < retryTimes; i++ {
-		isUpgrading, err := IsUpgradingClusterState(s)
-		if err == nil {
-			if isUpgrading {
-				break
-			}
-			logutil.BgLogger().Fatal("global state isn't upgrading, please send a request to start the upgrade first",
-				zap.String("category", "upgrading"), zap.Error(err))
-		}
-
-		if i == retryTimes-1 {
-			logutil.BgLogger().Fatal("get global state failed", zap.String("category", "upgrading"), zap.Error(err))
-		}
-		if i%10 == 0 {
-			logutil.BgLogger().Warn("get global state failed", zap.String("category", "upgrading"), zap.Error(err))
-		}
-		time.Sleep(interval)
-	}
-	logutil.BgLogger().Info("global state is upgrading", zap.String("category", "upgrading"),
-		zap.Int64("old version", ver), zap.Int64("latest version", currentBootstrapVersion))
 }
 
 // checkOwnerVersion is used to wait the DDL owner to be elected in the cluster and check it is the same version as this TiDB.
